vim: ku: Fix wrong regexp in auto directory completion
Fix the bug in s:text_by_auto_directory_completion().  It used
line_components[-2] as regular expression, but it may have meta characters, and
if so, it will cause some errors.  Fix this bug by using s:make_skip_regexp()
to escape meta characters in line_components[-2].

This change also revises the matching algorithm to determine which components
to be completed.  Originally, it should use skip matching, but it used a simple
algorithm which is easy to write in the early stage of the development of ku.