vim: ku: Fix to expand proepr prefix
Fix a bug in prefix expansion.  It didn't expand proper prefix if there are
2 or more prefixes which have the same head part.  For example, if there are
prefixes "ab", "abc", "abcd" and so forth, it is not possible to predict which
prefix will be expanded for user input "abcd/efgh".  It should expand "abcd",
but it may use "abc" or "ab" instead.

This change also includes caching result of s:prefix_table_for() in
s:expand_prefix3().