vim: vimrc: Fix {rhs}s to move in Insert mode
Use the cursor keys for mappings to move the cursor in Insert mode.
Old code uses <C-o>{command} but it doesn't work if the cursor is BEYOND
the end of the current line, because the cursor will be moved AT the
last character in the current line just after <C-o>.  As a result, the
cursor will be moved to unexpected position.