vim: ku: Fix the behavior of ku#restart()
The behavior of ku#restart() was changed since 0.1.6.
It might use a completed value instead of what user typed.
This behavior is incompatible with typical usage of ku#restart().

Fix this behavior to be the same as 0.1.6 - use what user typed.
Also add some notes for what value to be used.