vim: vimrc: ku: Fix timing to custom action/prefix
The old version delaied defining custom actions and prefixes for ku
until the ku buffer is created.  This was intended to speed up launching
a Vim process.

But it's somewhat troublesome to apply new definitions at runtime.
The new version stops to delay, because the time to launch a Vim process
with the new version is also very short and it can be ignored.

Note that the results of "time vim -c ':quit'" are:
Old version: 0.16s user 0.01s system 94% cpu 0.180 total
New version: 0.20s user 0.01s system 96% cpu 0.222 total