vim: syntaxcomplete: Fix wrong code on 'iskeyword'
s:SyntaxCSyntaxGroupItems() uses 'iskeyword' to extract keywords from an
output of :syntax list if b:omni_syntax_use_iskeyword is true.  But it
didn't parse 'iskeyword' correctly.  As a result, it returned the same
result regardless of the values of b:omni_syntax_use_iskeyword and
'iskeyword'.

Fix this bug with \k.
There is no reason to look into the complex 'iskeyword'.