vim: fakeclip: Fix the way to detect platform
Fix the way to detect the current platform, especially on Mac OS X.
It used the result of system('uname -o'), but the "-o" option is not
available for /usr/bin/uname and the "-o" option seems to be unnecessary.
I use GNU's uname instead, so that I couldn't find this bug.

Thanks for someone who wrote about this bug at the following page in the
Vim Tips Wiki: http://vim.wikia.com/wiki/Mac_OS_X_clipboard_sharing