vim: textobj-function: Rename and restructure
Rename the plugin name as textobj-function from tofunc, the old name.
Because tofunc is not clear to indicate that this plugin provides text
objects for functions.

Restructure names and locations of files.  Because the plugin name has
been changed, and there is no strong reason to put usage samples of
textobj-function in after/ftplugin - they just define several functions
and the function are not relvant to any buffer.