Makefile: _vimup: Fix the way to choose main doc
In _vimup, it assumed that the main document file of a package is listed
first among the documents in the package.  But now, vim-textobj-function
has 3 documents and the main document file is listed below, so that this
way is no more applicable.

Fix this problem with new assumption: The main document file must be
named as "vim/dot.vim/doc/{package name without the "vim-" prefix}.txt".