vim: surround: Fix a wrong link for this plugin
Recently, links to HTML-version documents of my Vim plugins were added
into the corresponding :help-version documents.  But I didn't care about
the document of "surround".  It is not written in the same format as
other plugins, so it should be left as-is, but I forgot about it.
Revert the change.