vim: ku: file: Fix bugs on dotfiles/suffixes/".."
On Microsoft Windows, some glob patterns are not expanded as same as
other environments especially *nix-like one.  For example, glob('*') and
glob('.*') are expanded as we expect, but glob('{,.}*') doesn't list
dotfiles on Microsoft Windowos.  Fix this bug by listing dotfiles and
others separately.

And another bug is also fixed by this change.  The bug is that
directories are suffixed with a path separator if
user_seems_want_dotfiles_p is true.  It is inconvenient with automatic
component completion.

Lastly, ".." was not listed by glob() implicitly, but ".." is now always
listed after this fix.  Remove ".." manually if user seems to find items
under the root directories.  Also improve the condition to determine
whether a given pattern means the root directory or not.