vim: ft issue: Fix a way to detect target to jump
To jump to the issue corresponding to an ID which is under or near the
cursor, search() is used with the flags 'cW'.  But with the flag 'cW',
jumping doesn't work as I expect if the cursor is located at the first
character of an ID.

Fix this problem by use the flags 'ceW' instead.