Fix the wrong meaning of root_directory_pattern_p
The meaning of _.root_directory_pattern_p returned by s:parse_pattern()
was changed between ku 0.2.3 and ku 0.2.4.  So it is not impossible to
list items in the most of absolute paths.  Fix the meaning of
_.root_directory_pattern_p to be the same as the old and the right one.

Anyway, the naming of "root_directory_pattern_p" seems to be easy to
misunderstand.  The naming should be fixed later.