vim: gtd: ftplugin: Fix a bug on b:undo_ftplugin
When b:undo_ftplugin is not defined, the filetype plugin set a value
starting with "|" to b:undo_ftplugin.  But it is wrong, because null
string may be treated as a command like :print in some cases.  For
example, What " | ..." does is the same as what "print | ..." does.

Fix this bug by removing "|" from b:undo_ftplugin without values of
other filetype plugins.