vim: fakeclip: Revise terms on GNU screen
To add support for tmux later, the current usage of terms on GNU screen
is not proper.  Revise usage of terms to express that fakeclip provides
a pseudo register for "paste buffer".

Key mappings such as <Plug>(fakeclip-screen-y) should be renamed, but
it's a troublesome work.  So I defer to do it.