vimrc: Treat <kEnter> as if it is a <Return>
One of my Kinesis Contoured Keyboards broke down; its Enter key (on the
top layer) does not respond anymore, so that I use another Enter key (on
the embedded layer) to map the function of Enter to other keys.

But the "embedded" Enter key is a keypad Enter key.  Some versions of
Vim distinguish <Enter> and <kEnter>.  I didn't consider such situation
when I wrote a setting on Return-Semicolon swapping.  So the setting
doesn't work comfortably with <kEnter>.

Fix this problem by treating <kEnter> as <Return>.