Change "d" to be a synonym for "diff"
I often used "git diff HEAD" to check both staged changes and unstaged
changes at the same time, so that I defined "d" to be a synonym for
"diff HEAD".  But I realized that the definition is not so good, because
I often forget about the implicit "HEAD" then I run into mysterious
problems.

To solve the above problem, change "d" to be a synonym for "diff", and
add "db" to do the same stuff as the old "d".