Add alias "derive" as alternative "checkout -b"
"checkout -b $topic $base" is not intuitive in a sense and it's not
convenient whenever I edit command line, because I have to type
"<Esc>/checkout -b<Esc>$Bi<C-w>{new-topic}".
                       ^^^^^^^^

Add alias "derive" (usage: "derive $base $topic") which is expanded to
"checkout -b $topic $base".  With this alias, the number of keys to type
is reduced as follows: "<Esc>/derive<Esc>A<C-w>{new-topic}".
                                         ^^^^^^