vimrc: VCS branch name: Improve appearance
(a) "|" is used to separate branch_name and additional_info.  But "|" is
already used to separate names of tabpages and a VCS branch name.
Change display about additional_info to avoid confusion.

(b) Change alternative label for non-git repository to avoid confusion
with a case that the current branch name is not known.