<Esc>[ causes a catastrophe in terminal
Many escape sequences start with <Esc>[.  Especially, if a response to
a query starts with <Esc>[, the response will be treated as a part of
user input.  For example, <Esc>[6n (query the cursor position) causes
a terminal to respond with <Esc>[{ROW};{COL}R -- and this response
results a random stuff.